+++ spec/lib/slkecho/slack_request/post_message_spec.rb
@@ -13,11 +13,12 @@ RSpec.describe Slkecho::SlackRequest::PostMessage do
 
     before do
       stub_request(:post, "https://slack.com/api/chat.postMessage")
-        .to_return(status: status_code, body: response_body)
+        .to_return(status: [status_code, status_message], body: response_body)
     end
 
     context "when API request is successful" do
       let(:status_code) { 200 }
+      let(:status_message) { "OK" }
       let(:response_body) { { ok: true }.to_json }
 
       it { is_expected.to be_truthy }
@@ -25,6 +26,7 @@ RSpec.describe Slkecho::SlackRequest::PostMessage do
 
     context "when API request is not successful" do
       let(:status_code) { 200 }
+      let(:status_message) { "OK" }
       let(:response_body) { { ok: false, "error": "too_many_attachments" }.to_json }
 
       it "raises SlackApiResultError" do
@@ -34,10 +36,11 @@ RSpec.describe Slkecho::SlackRequest::PostMessage do
 
     context "when HTTP error respond" do
       let(:status_code) { 400 }
-      let(:response_body) { "Bad Request" }
+      let(:status_message) { "Bad Request" }
+      let(:response_body) { "" }
 
       it "raises SlackApiHttpError" do
-        expect { subject }.to raise_error(Slkecho::SlackApiHttpError, "Bad Request")
+        expect { subject }.to raise_error(Slkecho::SlackApiHttpError, "400 Bad Request")
       end
     end
   end