+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: Rename errors
   - `Slkecho::SlackApiRequestError` to `Slkecho::SlackApiHttpError`
   - `Slkecho::SlackApiResponseErorr` to `Slkecho::SlackApiResultError`
+- feat: Improve error message of Slack API request
 
 ## [v1.4.0] - 2024-01-09